Helper08/htmlreorg (#80)
* Started finishing up the code for skipping MABs

* Finished the MAB exclusion thing

* Fixed attendance to only track brothers who were active when the event
was created. Created methods to check attendance ratios that can be
called on in template rather than passing these as context. Made events
use multi-table inheritance rather than abstract class inheritance,
meaning excuses can now be made for all events, not just chapter events

* Added a way for chairs and leadership to add brothers to the list of
eligible_attendees after the fact. When brothers are added to committees
they're automatically added to committee events as eligible_attendees if
the commitee event has not occurred yet. If they're removed from a
committee they are taken off of the list of eligible_attendees for
future committee meetings.

* Made the committee assingment function more efficient

* added comments

* Changed one of the datetime.dates to datetime.datetime

* Added an excuse form to other events and added better UI for resolving
and reading excuses for the secretary. Additionally, the recruitment
chair will be able to read and address recruitment events as well.

* Finished up everything related to excuses, brothers can now write and
track excuses for all events (with the exception of study_table events
and committee meetings). Additionally, brothers can easily manage and
track their attendance for each type of event through a fraction at the
bottom of each panel on the brother page.

* Added comments to multiple views and methods, generalized some redundant
views

* Cleaned up somoe views and forms

* Update dashboard/forms.py

Co-authored-by: Jared Cassarly <jaredcassarly@case.edu>

* Addressed some minor issues

* Added documentation and resolved redundancies

* Bug fixes

* Finished replacing code with a util function

* Edited some of the templates to be similar to each other

* Did some funky stuff with html files

* In the process of editing html reorg

* Cleaned up templates, renamed templates that inherit from other
templates, fixed inheritance types for some templates.
Used inheritance where html was repeated extensively
Created a tag that adds an asterisk to a field lable if it is required
Created a tag directory and a custom_tag file to load the tag that was
created.

* Changed around html and cleaned up some redundant code`

* Adding additional documentation and cleaning up imports and variable names

* Adding additional documentation and cleaning up imports and variable names. Fixed minor bugs and cleaned up verify_positions. Strings that represent position titles were changed to be slugs to match the new title scheme.

* Cleaned up some views, imports and html

* capitalized constant names and fixed a docstring

* added docstring for template tag

* added docstring to DashboardUpdateView and DashboardDeleteView

Co-authored-by: Jared Cassarly <jaredcassarly@case.edu>